Fix several AM62 tests

+++ tests/suites/am62/graphics/setup_suite.bash
@@ -15,12 +15,18 @@ setup_suite() {
 
 teardown_suite() {
     docker container stop graphics-tests
-    docker image rm -f $(docker container inspect -f '{{.Image}}' graphics-tests)
+
+    if [ -z "$DO_NOT_RM_ON_TEARDOWN" ]; then
+        docker image rm -f $(docker container inspect -f '{{.Image}}' graphics-tests)
+    else
+        echo "Skipping Docker image removal due to DO_NOT_RM_ON_TEARDOWN environment variable."
+    fi
+
     docker container rm graphics-tests
 
     for dir in /sys/class/drm/card*-HDMI-*; do
         if [[ -d $dir ]]; then
             echo "off" > "$dir/status"
-         fi
+        fi
     done
 }
